+++ zulu/__pkg__.py
@@ -6,7 +6,7 @@ __package_name__ = 'zulu'
 __description__ = ('A drop-in replacement for native datetimes '
                    'that embraces UTC')
 __url__ = 'https://github.com/dgilland/zulu'
-__version__ = '0.6.0'
+__version__ = '0.7.0'
 __author__ = 'Derrick Gilland'
 __email__ = 'dgilland@gmail.com'
 __license__ = 'MIT License'